refactor: ViewResolver 개선
