+++ src/main/java/model/dto/user/UserListViewDto.java
@@ -1,4 +1,4 @@
-package model.dto;
+package model.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
