+++ src/main/java/utils/builder/RequestBuilder.java
@@ -1,7 +1,11 @@
-package utils;
+package utils.builder;
 
 import lombok.experimental.UtilityClass;
+import model.request.properties.HttpRequestFirstLineProperties;
 import model.request.*;
+import model.dto.request.HttpRequestKeyValue;
+import model.request.properties.RequestBody;
+import model.request.properties.RequestHeaders;
 
 import java.io.BufferedReader;
 import java.io.IOException;
