+++ src/main/java/utils/extractor/RequestBodyExtractor.java
@@ -1,20 +1,21 @@
-package model.request;
+package utils.extractor;
 
 import com.fasterxml.jackson.core.type.TypeReference;
+import lombok.experimental.UtilityClass;
 import model.enumeration.ContentType;
-import utils.ObjectMapperFactory;
+import utils.factory.ObjectMapperFactory;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.Map;
 
 import static constant.HeaderConstant.*;
-import static utils.IOUtils.*;
-import static utils.QueryStringParser.*;
+import static utils.utils.IOUtils.*;
+import static utils.parser.QueryStringParser.*;
 
+@UtilityClass
 public class RequestBodyExtractor {
-    public static Map<String, String> extract(Map<String, String> requestHeaders, BufferedReader bufferedReader) {
-        // todo: 와 여기 어떻게 못하나 ㅋㅋ
+    public Map<String, String> extract(Map<String, String> requestHeaders, BufferedReader bufferedReader) {
         try {
             switch (ContentType.of(requestHeaders.get(CONTENT_TYPE))) {
                 case APPLICATION_JSON:
