+++ src/main/java/utils/utils/FileIoUtils.java
@@ -1,11 +1,10 @@
-package utils;
+package utils.utils;
 
 import lombok.experimental.UtilityClass;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
-import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
