+++ src/main/java/utils/utils/IOUtils.java
@@ -1,4 +1,4 @@
-package utils;
+package utils.utils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
