+++ src/test/java/utils/FileIoUtilsTest.java
@@ -3,6 +3,7 @@ package utils;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import utils.utils.FileIoUtils;
 
 public class FileIoUtilsTest {
     private static final Logger log = LoggerFactory.getLogger(FileIoUtilsTest.class);
