+++ src/test/java/webserver/RequestHandlerTest.java
@@ -4,7 +4,7 @@ import db.DataBase;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import support.StubSocket;
-import utils.FileIoUtils;
+import utils.utils.FileIoUtils;
 import webserver.infra.RequestHandler;
 
 import java.io.IOException;
